Move enum outside model
