+++ megatron/arguments.py
@@ -21,7 +21,7 @@ import os
 import torch
 import deepspeed
 
-from megatron.model.enums import PositionEmbeddingType
+from megatron.enums import PositionEmbeddingType
 
 
 def parse_args(extra_args_provider=None, defaults={},
@@ -308,7 +308,7 @@ def _add_network_size_args(parser):
     group.add_argument('--bert-no-binary-head', action='store_false',
                        help='Disable BERT binary head.',
                        dest='bert_binary_head')
-    group.add_argument('--position-embedding-type', type=PositionEmbeddingType, choices=list(PositionEmbeddingType),
+    group.add_argument('--position-embedding-type', type=lambda x: PositionEmbeddingType[x], choices=list(PositionEmbeddingType),
                        help='Define position embedding type.'
                        )
