+++ megatron/data/Makefile
@@ -1,5 +1,5 @@
 CXXFLAGS += -O3 -Wall -shared -std=c++11 -fPIC -fdiagnostics-color
-CPPFLAGS += $(shell python3 -m pybind11 --includes)
+CPPFLAGS += $(shell python -m pybind11 --includes)
 LIBNAME = helpers
 LIBEXT = $(shell python3-config --extension-suffix)
